==> Codes/1.ipynb <==
#Unique Entities and Possible Combination of Two Related Entities
**`Input`**`:-`<br>
1. `process_flow_sentences.csv`<br>
**`Output`**`:-`<br>
1. `unique_entities.csv: /content/drive/MyDrive/Final/unique_entities.csv`
2. `all_entities_pair.csv: /content/drive/MyDrive/Final/all_entities_pair.csv`

import pandas as pd
from google.colab import drive 
drive.mount('/content/drive') 

pfs=pd.read_csv('/content/drive/MyDrive/Final (Completed)/process_flow_sentences_2.csv')
pfs

#Get All the Unique Entities from process_flow_sentences.csv
ent1=(pfs['ent_1'].drop_duplicates()).reset_index(drop=True)
ent2=(pfs['ent_2'].drop_duplicates()).reset_index(drop=True)
ent=[]
for e in ent1:
    ent.append(e)
for e in ent2:
    ent.append(e)
ent=list(set(ent))
ent

ue=pd.DataFrame({'unique entities':ent})
ue

ue.to_csv('/content/drive/MyDrive/Final (Completed)/unique_entities.csv',index=False)

#Just getting all entities pairs from process flow sentences
ent1=pfs['ent_1'].reset_index(drop=True)
ent2=pfs['ent_2'].reset_index(drop=True)

#All Entities Pair - aep
aep=pd.DataFrame({'ent1':ent1,'ent2':ent2}) 
print(len(aep))
aep

aep.to_csv('/content/drive/MyDrive/Final (Completed)/all_entities_pair.csv',index=False)

#End Of File
