==> Codes/6.ipynb <==
#**6:**
**1. Entity-Relation Extraction**<br>
**2. Creation of Triples for KG along with properties of Nodes/Entities**<br>
`Input: process_flow_sentneces.csv, PMJAY_NLP_Model`<br>
`Output: triples.csv or triple.json`


from google.colab import drive
drive.mount('/content/drive/')

import pandas as pd
pfs= pd.read_csv('/content/drive/MyDrive/Final (Completed)/process_flow_sentences_2.csv')

pfs

sentences=pfs['sentences']
sentences

## Extration of Entities

import spacy
pmjay_nlp= spacy.load('/content/drive/MyDrive/Final (Completed)/nlp_model_2')

#Extration of Entities from pfs sentences
#Create entity-pair dictionary and ent-label dictionary

ent_pair=[] # [(sentence,[ent_1, ent_2]),...]
el_dict={} #Entity-Label dictionary
for i in range(0,len(sentences)):
    e_pair=[]
    d= pmjay_nlp(sentences[i])
    for ent in d.ents:        
        e_pair.append(ent)
        el_dict[ent.text]=ent.label_
    if(len(e_pair)>1):
        ent_pair.append((sentences[i],e_pair))
print("Entity Pairs:",len(ent_pair),"\n",ent_pair)
print("Entity Label Dict:",len(el_dict),"\n",el_dict)

#Adding Missing Data
el_dict['Am I Eligible App']='MOBILE_APP'
el_dict['Am I Eligible Website']='PORTAL'
el_dict['Medication']='PROCESS'
el_dict['Hospitalization']='PROCESS'
ent_pair.append((("The Patient checks eligibility through Am I Eligible App"),['Patient','Am I Eligible App']))
ent_pair.append((("The Patient checks eligibility through Am I Eligible Website"),['Patient','Am I Eligible Website']))
ent_pair.append((("General Doctor decides Hospitalization"),['General Doctor','Hospitalization']))
ent_pair.append((("General Specialist decides Hospitalization"),['General Specialist','Hospitalization']))
ent_pair.append((("General Doctor decides Medication"),['General Doctor','Medication']))
ent_pair.append((("General Specialist decides Medication"),['General Specialist','Medication']))
print("Entity Pairs:",len(ent_pair),"\n",ent_pair)
print("Entity Label Dict:",len(el_dict),"\n",el_dict)

#Extraction of Relations and creating Level 1 triplets for KG
import re
triplets_1=[] # [[ent_1, ent_2, relation],...]
for e in ent_pair:
    e1= str(e[1][0])
    e2= str(e[1][1])
    sent= e[0]
    result= re.search(e1+"(.*)"+e2,sent)
    triplets_1.append([e1,e2,result.group(1).strip()])
triplets_1 

trip_df=pd.DataFrame(triplets_1)
trip_df.rename(columns={0:'source',1:'target',2:'edge'},inplace=True)
trip_df

#trip_df.to_csv("/content/drive/MyDrive/Final (Completed)/triplets.csv",index=False)

# Adding Properties to unique Entities
ue= pd.read_csv('/content/drive/MyDrive/Final (Completed)/unique_entities_added_properties.csv')
#ue['2']

ue.sample(5)

#Getting Properties for each Entity
ep_dict={} #dictionary containing entity:properties
ue_2=ue.drop(['0'], axis=1)
for i in range(0,len(ue_2)):
    d=(ue_2.iloc[i])
    prop=""
    for v in d:
        #print(v,":",type(v))
        if(type(v)==type('')):
            prop=prop+" "+str(v)+","
    prop=(prop[:-1]).strip()
    ep_dict[ue['0'][i]]=prop 
ep_dict

#assign unique variables to each entity 
#maintain a dict of {entity:variable} to use it for creating relations
def generate_variable(n):
    alphabet = "abcdefghijklmnopqrstuvwxyz"
    sequences = []
    if n < 26:
        return(alphabet[n])
    else:
        first = alphabet[(n // 26) - 1]
        second = alphabet[n % 26]
        return(first + second)
    return sequences
ev_dict={}
for i in range(0,len(ue)):
    var=generate_variable(i)
    e=str(ue['0'][i])
    ev_dict[e]=var 

#Cypher Query for Creating unique nodes with properties
c_query=[]
for i in range(0,len(ue)):
    en=ue['0'][i]
    prop=ue['1'][i]
    q="CREATE ("+ev_dict[en]+":"+el_dict[en]+"{"+ep_dict[en]+"})"
    c_query.append(q)
c_query

#Use rel_pair to make relations cypher query
rel_pairs=[] #(source entity's variable, relation, target entitiy's variable)
for r in triplets_1:
    rel_pairs.append((ev_dict[r[0]],r[2],ev_dict[r[1]]))
rel_pairs 

#Cypher Query for Creating Relations between entities
# CREATE (a)-[:relation]->(b)
#import re
for r in rel_pairs:
    rel= re.sub(' ','_',r[1])
    c_query.append("CREATE ("+r[0]+")-[:"+rel+"]->("+r[2]+")")

for q in c_query:
    print(q)

'''for r in rel_pairs:
    rel= re.sub(' ','_',r[1])
    c_query.append("CREATE ("+r[0]+")-[:"+rel+"]->("+r[2]+")")
cypher_text=""
for q in c_query:
    cypher_text=cypher_text+q+"\n"
cypher_text=cypher_text.strip()'''

cypher_text=""
for q in c_query:
    cypher_text=cypher_text+q+"\n"
cypher_text=cypher_text.strip()
print(cypher_text)

text_file = open("/content/drive/MyDrive/Final (Completed)/pmjay_process_flow.cypher", "w")
n = text_file.write(cypher_text)
text_file.close()

# End Of file
